refactor: rename package to com.accelerex.swagger
